requirements.txt
fastapi
uvicorn
ultralytics
opencv-python-headless
numpy
pillow
python-multipart
